+++ demisto_sdk/validation/file_validator.py
@@ -21,7 +21,7 @@ from demisto_sdk.common.constants import CODE_FILES_REGEX, OLD_YML_FORMAT_FILE, 
     IMAGE_REGEX, TEST_PLAYBOOK_REGEX, INTEGRATION_YML_REGEX, DIR_LIST_FOR_REGULAR_ENTETIES, \
     PACKAGE_SUPPORTING_DIRECTORIES, YML_BETA_INTEGRATIONS_REGEXES, PACKAGE_SCRIPTS_REGEXES, YML_INTEGRATION_REGEXES, \
     PACKS_DIR, PACKS_DIRECTORIES, Errors, PLAYBOOKS_REGEXES_LIST, JSON_INDICATOR_AND_INCIDENT_FIELDS, PLAYBOOK_REGEX, \
-    JSON_ALL_LAYOUT_REGEXES
+    JSON_ALL_LAYOUT_REGEXES, REPUTATION_REGEX, CHECKED_TYPES_REGEXES
 from demisto_sdk.common.hook_validations.conf_json import ConfJsonValidator
 from demisto_sdk.common.hook_validations.description import DescriptionValidator
 from demisto_sdk.common.hook_validations.id import IDSetValidator
@@ -46,21 +46,25 @@ class FilesValidator:
     to make sure you did not bypass the hooks as a safety precaution.
 
     Attributes:
-        _is_valid (bool): saves the status of the whole validation(instead of mingling it between all the functions).
-        conf_json_validator (ConfJsonValidator): object for validating the conf.json file.
-        id_set_validator (IDSetValidator): object for validating the id_set.json file(Created in Circle only).
+        is_backward_check (bool): Whether to check for backwards compatibility.
+        prev_ver (str): If using git, holds the branch to compare the current one to. Default is origin/master.
+        use_git (bool): Whether to use git or not.
+        is_circle: (bool): Whether the validation was initiated by CircleCI or not.
+        print_ignored_files (bool): Whether to print the files that were ignored during the validation or not.
+        validate_conf_json (bool): Whether to validate conf.json or not.
+        validate_id_set (bool): Whether to validate id_set or not.
+        file_path (string): If validating a specific file, golds it's path.
+        configuration (Configuration): Configurations for IDSetValidator.
     """
 
     def __init__(self, is_backward_check=True, prev_ver='origin/master', use_git=False, is_circle=False,
-                 print_ignored_files=False, validate_conf_json=True, validate_id_set=False,
+                 print_ignored_files=False, validate_conf_json=True, validate_id_set=False, file_path=None,
                  configuration=Configuration()):
         self.branch_name = ''
         self.use_git = use_git
         if self.use_git:
             print('Using git')
-            branches = run_command('git branch')
-            branch_name_reg = re.search(r'\* (.*)', branches)
-            self.branch_name = branch_name_reg.group(1)
+            self.branch_name = self.get_current_working_branch()
             print(f'Running validation on branch {self.branch_name}')
 
         self.prev_ver = prev_ver
@@ -75,6 +79,7 @@ class FilesValidator:
         self.print_ignored_files = print_ignored_files
         self.validate_conf_json = validate_conf_json
         self.validate_id_set = validate_id_set
+        self.file_path = file_path
 
         if self.validate_conf_json:
             self.conf_json_validator = ConfJsonValidator()
@@ -89,6 +94,12 @@ class FilesValidator:
         else:
             print_color('The files were found as invalid, the exact error message can be located above', LOG_COLORS.RED)
             return 1
+
+    @staticmethod
+    def get_current_working_branch():
+        branches = run_command('git branch')
+        branch_name_reg = re.search(r'\* (.*)', branches)
+        return branch_name_reg.group(1)
 
     @staticmethod
     def get_modified_files(files_string, tag='master', print_ignored_files=False):
@@ -176,6 +187,7 @@ class FilesValidator:
             'git diff --name-status {tag}..{compare_type}refs/heads/{branch}'.format(tag=tag,
                                                                                      branch=self.branch_name,
                                                                                      compare_type=compare_type))
+
         modified_files, added_files, _, old_format_files = self.get_modified_files(
             all_changed_files_string,
             tag=tag,
@@ -190,6 +202,15 @@ class FilesValidator:
             modified_files_from_tag, added_files_from_tag, _, _ = \
                 self.get_modified_files(all_changed_files_string,
                                         print_ignored_files=self.print_ignored_files)
+
+            if self.file_path:
+                if F'M\t{self.file_path}' in files_string:
+                    modified_files = {self.file_path}
+                    added_files = set()
+                else:
+                    modified_files = set()
+                    added_files = {self.file_path}
+                return modified_files, added_files, set(), set()
 
             old_format_files = old_format_files.union(nc_old_format_files)
             modified_files = modified_files.union(
@@ -326,6 +347,14 @@ class FilesValidator:
             elif 'CHANGELOG' in file_path:
                 self.is_valid_release_notes(file_path)
 
+            elif checked_type(file_path, [REPUTATION_REGEX]):
+                print_color(
+                    F'Skipping validation for file {file_path} since no validation is currently defined.',
+                    LOG_COLORS.YELLOW)
+
+            elif checked_type(file_path, CHECKED_TYPES_REGEXES):
+                pass
+
             else:
                 print_error("The file type of {} is not supported in validate command".format(file_path))
                 print_error("'validate' command supports: Integrations, Scripts, Playbooks, "
@@ -417,6 +446,11 @@ class FilesValidator:
             elif 'CHANGELOG' in file_path:
                 self.is_valid_release_notes(file_path)
 
+            elif checked_type(file_path, [REPUTATION_REGEX]):
+                print_color(
+                    F'Skipping validation for file {file_path} since no validation is currently defined.',
+                    LOG_COLORS.YELLOW)
+
             else:
                 print_error("The file type of {} is not supported in validate command".format(file_path))
                 print_error("validate command supports: Integrations, Scripts, Playbooks, "
@@ -441,9 +475,7 @@ class FilesValidator:
             self._is_valid = False
 
     def validate_committed_files(self):
-        """Validate that all the committed files in your branch are valid
-
-        """
+        """Validate that all the committed files in your branch are valid"""
         modified_files, added_files, old_format_files, packs = self.get_modified_and_added_files()
         schema_changed = False
         for f in modified_files:
@@ -531,16 +563,20 @@ class FilesValidator:
         if self.use_git:
             if self.branch_name != 'master' and (not self.branch_name.startswith('19.') and
                                                  not self.branch_name.startswith('20.')):
-                print("Validates only committed files")
+                print('Validates only committed files')
                 self.validate_committed_files()
                 self.validate_against_previous_version(no_error=True)
             else:
                 self.validate_against_previous_version(no_error=True)
-                print("Validates all of Content repo directories according to their schemas")
+                print('Validates all of Content repo directories according to their schemas')
                 self.validate_all_files()
         else:
-            print("No using git, validating all files")
-            self.validate_all_files()
+            if self.file_path:
+                self.branch_name = self.get_current_working_branch()
+                self.validate_committed_files()
+            else:
+                print('Not using git, validating all files')
+                self.validate_all_files()
 
         return self._is_valid